Update README.md
Updated the README file to include more information about the web app as well as fixed the rendering of the screenshots in the README. Included a couple sections, included links to our sources and some formatting changes.

After deployment, we need to take a screenshot of the deployed page in the browser. I set the filename of that screenshot in the README to project1live.png

All we need to do is put the project1live.png file into the images folder and link the deployed website in the README file and it will be completed.